refactor(python): split bindings into private modules
