+++ python/dataprof/_profiler.pyi
@@ -1,0 +1,42 @@
+"""Type stubs for _profiler."""
+
+from __future__ import annotations
+
+from collections.abc import Callable
+from typing import Any
+
+from ._dataprof import ProgressEvent, SamplingStrategy, StopCondition
+from ._report import ProfileReport
+
+class Profiler:
+    """Builder-style profiler configuration.
+
+    Chainable methods accumulate settings; call ``.profile(source)`` to run.
+
+    Example::
+
+        report = dp.Profiler().engine("incremental").max_rows(5000).profile("data.csv")
+    """
+
+    def __init__(self) -> None: ...
+    def engine(self, engine: str) -> Profiler: ...
+    def chunk_size(self, n: int) -> Profiler: ...
+    def memory_limit_mb(self, mb: int) -> Profiler: ...
+    def format(self, fmt: str) -> Profiler: ...
+    def max_rows(self, n: int) -> Profiler: ...
+    def name(self, name: str) -> Profiler: ...
+    def csv_delimiter(self, d: str) -> Profiler: ...
+    def csv_flexible(self, flexible: bool) -> Profiler: ...
+    def sampling(self, strategy: SamplingStrategy) -> Profiler: ...
+    def stop_condition(self, cond: StopCondition) -> Profiler: ...
+    def on_progress(self, cb: Callable[[ProgressEvent], None]) -> Profiler: ...
+    def progress_interval_ms(self, ms: int) -> Profiler: ...
+    def quality_dimensions(self, dims: list[str]) -> Profiler: ...
+    def columns(self, columns: list[str]) -> Profiler: ...
+    def stop_when(self, condition: StopCondition | str) -> Profiler: ...
+    def metrics(self, packs: list[str]) -> Profiler: ...
+    def locale(self, locale: str) -> Profiler: ...
+    def positive_columns(self, columns: list[str]) -> Profiler: ...
+    def identifier_columns(self, columns: list[str]) -> Profiler: ...
+    def temporal_columns(self, columns: list[str]) -> Profiler: ...
+    def profile(self, source: Any) -> ProfileReport: ...
