+++ python/dataprof/_report.pyi
@@ -1,0 +1,121 @@
+"""Type stubs for _report."""
+
+from __future__ import annotations
+
+from collections.abc import Iterator
+from os import PathLike
+from typing import Any
+
+from ._dataprof import ColumnProfile, DataQualityMetrics
+
+class ProfileReport:
+    """High-level profiling report with export methods.
+
+    Wraps the raw ``dataprof._dataprof.ProfileReport`` (or a dict-backed proxy
+    for reloaded reports). Supports dict-like column access::
+
+        report["column_name"]          # -> ColumnProfile
+        "column_name" in report        # -> bool
+        for name in report: ...        # iterate column names
+        len(report)                    # number of columns
+    """
+
+    def __init__(self, report: Any) -> None: ...
+    @property
+    def source(self) -> str: ...
+    @property
+    def source_type(self) -> str: ...
+    @property
+    def engine(self) -> str | None: ...
+    @property
+    def rows(self) -> int: ...
+    @property
+    def columns(self) -> int: ...
+    @property
+    def column_profiles(self) -> dict[str, ColumnProfile]: ...
+    @property
+    def profiles(self) -> list[ColumnProfile]: ...
+    @property
+    def quality_score(self) -> float | None: ...
+    @property
+    def quality(self) -> DataQualityMetrics | None: ...
+    @property
+    def semantic_hint_bindings(self) -> list[dict[str, Any]]: ...
+    @property
+    def low_sample_warning(self) -> bool: ...
+    @property
+    def execution_time_ms(self) -> int: ...
+    @property
+    def throughput(self) -> float | None: ...
+    @property
+    def memory_peak_mb(self) -> float | None: ...
+    @property
+    def truncation_reason(self) -> str | None: ...
+    @property
+    def source_exhausted(self) -> bool: ...
+    @property
+    def error_count(self) -> int: ...
+    @property
+    def ragged_row_count(self) -> int: ...
+    @property
+    def sampling_applied(self) -> bool: ...
+    @property
+    def sampling_ratio(self) -> float | None: ...
+
+    # Mapping protocol
+    def __getitem__(self, key: str) -> ColumnProfile: ...
+    def __contains__(self, key: object) -> bool: ...
+    def __iter__(self) -> Iterator[str]: ...
+    def __len__(self) -> int: ...
+
+    # Export methods
+    def to_dict(self) -> dict[str, Any]: ...
+    def to_json(self, indent: int = 2) -> str: ...
+    def to_dataframe(self) -> Any:
+        """Column profiles as pandas DataFrame (all stats, rounded). Requires pandas."""
+        ...
+    def to_polars(self) -> Any:
+        """Column profiles as polars DataFrame (all stats, rounded). Requires polars."""
+        ...
+    def to_arrow(self) -> Any:
+        """Column profiles as PyArrow Table (all stats, rounded). Requires pyarrow."""
+        ...
+    def describe(self) -> Any:
+        """Transposed summary (like pandas describe). Returns pandas DataFrame or dict."""
+        ...
+    def quality_summary(self) -> dict[str, Any]:
+        """Single-row quality summary dict for easy aggregation."""
+        ...
+    def save(self, path: str | PathLike[str]) -> ProfileReport:
+        """Save report to file (.json, .csv, or .parquet), returning self for chaining."""
+        ...
+    def to_html(self) -> str:
+        """Standalone HTML representation (same as Jupyter's rich display)."""
+        ...
+    def to_markdown(self) -> str:
+        """GitHub-flavored markdown table of the column profiles."""
+        ...
+    def to_llm_context(self, max_tokens: int = 1000, include_samples: bool = False) -> str:
+        """Token-bounded, agent-oriented summary: shape, caveats, flags, schema, patterns.
+
+        Tokens are estimated as ``ceil(len(text) / 4)``, not counted with a real
+        tokenizer. Raw cell values are excluded by default; ``include_samples=True``
+        only includes non-sensitive numeric extrema.
+        """
+        ...
+    def _to_llm_context(self, *, source: str, max_tokens: int, include_samples: bool) -> str: ...
+    def compare(self, other: ProfileReport) -> dict[str, Any]:
+        """Dict of quality/schema/null deltas versus another report."""
+        ...
+    @classmethod
+    def from_dict(cls, data: dict[str, Any]) -> ProfileReport:
+        """Rebuild a read-only ProfileReport from a to_dict() mapping."""
+        ...
+    @classmethod
+    def from_json(cls, text: str) -> ProfileReport:
+        """Rebuild a read-only ProfileReport from a to_json() string."""
+        ...
+    @classmethod
+    def load(cls, path: str | PathLike[str]) -> ProfileReport:
+        """Reload a report from a .json file written by save()."""
+        ...
