+++ python/dataprof/_report_schema.pyi
@@ -1,0 +1,7 @@
+"""Type stubs for _report_schema."""
+
+from __future__ import annotations
+
+REPORT_SCHEMA_VERSION: int
+
+_QUALITY_DIMENSIONS: tuple[str, ...]
